feat: moves user operations to a nested User struct
